Struct bevy_rrise::emitter_listener::RrEmitter

pub struct RrEmitter {
    pub event_id: AkID<'static>,
    pub flags: AkCallbackType,
    pub auto_post: bool,
    pub despawn_on_silent: bool,
    /* private fields */
}

Sound emitter configuration.

If its entity gets destroyed or this component gets removed, the events posted with it will be stopped.

Fields
event_id: AkID<'static>

The event to pre-set on this emitter. Defaults to no event (ie, "").

See auto_post

flags: AkCallbackType

Mask describing which callbacks you want to subscribe to. Defaults to none (ie, AkCallbackType(0)).

auto_post: bool

Whether to auto post the associated event when this emitter gets registered.

See event_id

despawn_on_silent: bool

Whether to automatically despawn the entity bearing this emitter when it is done playing.

Remark “Done playing” = no more events are playing on it - work if several events got posted simultaneously with it.

Implementations

impl RrEmitter

pub fn is_playing(&self) -> bool

Whether any events are playing on this emitter

pub fn is_registered(&self) -> bool

Whether this component appears to be registered in Wwise.

You can make sure of this by also querying for the RrRegistered component on your entities.

pub fn stop(&self)

Stops all events currently playing on this emitter.

pub fn post_associated_event(
&mut self,
cb_channel: Option<CallbackChannel>
) -> AkPlayingID

Posts the event self.event_id using flags self.flags.

If you pass None for cb_channel, you won’t receive any AkCallbackEvent in your [EventReader]s, even if you had some flags set in self.flags.

See CallbackChannel

pub fn post_event<'b, T: Into<AkID<'b>>>(
&mut self,
event: T,
flags: AkCallbackType,
cb_channel: Option<CallbackChannel>
) -> AkPlayingID

Posts event using flags (this method ignores self.flags).

If you pass None for cb_channel, you won’t receive any AkCallbackEvent in your [EventReader]s, even if you had some flags.

See CallbackChannel