Struct bevy_rrise::emitter_listener::RrListenerBundle

pub struct RrListenerBundle {
    pub tfm: TransformBundle,
    pub listener: RrListener,
}

Sound listener.

You should attach this to a camera or your player avatar.

Example

You can create a system that attaches any newly created listeners to your main 3D camera:

use bevy::prelude::*;
use bevy::render::camera::Camera3d;
use bevy_rrise::emitter_listener::RrListener;
fn attach_default_listeners_to_camera(
    mut cmds: Commands,
    listeners: Query<(Entity, &RrListener), Added<RrListener>>,
    main_camera: Query<Entity, With<Camera3d>>,
) {
    let main_camera = main_camera.single();
    for (entity, listener) in listeners.iter() {
        cmds.entity(main_camera).add_child(entity);
    }
}

Fields
tfm: TransformBundlelistener: RrListener
Implementations

impl RrListenerBundle

pub fn with_is_default(self, is_default: bool) -> Self

Sets whether this listener is a default listener or not.

Emitters that have not explicitly overridden their listener set are associated to the default listeners set.

Defaults to true.

impl RrListenerBundle

pub fn new(position: Vec3) -> Self

Creates a listener at position.

pub fn with_rotation(self, rotation: Quat) -> Self

Sets the rotation of this listener.